refactor(catalog): modularize site generation
