+++ examples/recipes/nvidia/video-search-and-summarization/README.md
@@ -137,25 +137,25 @@ this page. No VSS deployment, sandbox, notebook, or GPU host was involved.
 
 ```text
 $ python3 scripts/build_catalog.py --validate-metadata
-README catalog metadata is valid for 18 examples.
+README catalog metadata is valid for 19 examples.
 
 $ python3 scripts/build_catalog.py --write
-Built 18 catalog entries and detail pages in _site.
+Built 19 catalog entries and detail pages in _site.
 
 $ python3 scripts/build_catalog.py --check
-Catalog metadata and generated sources are valid: 18 examples across 7 browse groups.
+Catalog metadata and generated sources are valid: 19 examples across 7 browse groups.
 
-$ python3 -m unittest discover -s scripts/tests -p 'test_build_catalog.py'
-Ran 32 tests in 0.928s
+$ python3 -m unittest discover -s scripts/tests -p 'test_*.py'
+Ran 67 tests
 
 OK
 
-$ node --test scripts/tests/catalog.test.mjs
-# pass 12
+$ node --test scripts/tests/*.test.mjs
+# pass 14
 # fail 0
 
 $ python3 scripts/check_license_headers.py --check
-All 608 checked source files have SPDX headers.
+All 630 checked source files have SPDX headers.
 
 $ git diff --check
 ```
