+++ site/templates/detail.html
@@ -13,6 +13,8 @@
     <title>{{PAGE_TITLE}} | NemoClaw Community</title>
     <link rel="icon" type="image/png" sizes="64x64" href="{{FAVICON_URL}}">
     <link rel="stylesheet" href="{{STYLES_URL}}">
+    <link rel="stylesheet" href="{{DETAIL_STYLES_URL}}">
+{{PAGE_STYLES}}
   </head>
   <body class="detail-page{{PAGE_CLASS}}">
     <a class="skip-link" href="#readme">Skip to content</a>
