+++ site/templates/index.html
@@ -13,7 +13,8 @@
     >
     <title>NemoClaw Community Example Catalog | NVIDIA</title>
     <link rel="icon" type="image/png" sizes="64x64" href="assets/nvidia-favicon.png">
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="styles/shared.css">
+    <link rel="stylesheet" href="styles/catalog.css">
   </head>
   <body>
     <a class="skip-link" href="#catalog">Skip to examples</a>
@@ -252,6 +253,6 @@
       </div>
     </footer>
 
-    <script type="module" src="catalog.mjs"></script>
+    <script type="module" src="scripts/catalog.mjs"></script>
   </body>
 </html>
